feat/SSM-114: split queue logic from index_controller
